+++ src/main/java/com/scubakay/autorelog/commands/AutoRelogCommand.java
@@ -2,6 +2,7 @@ package com.scubakay.autorelog.commands;
 
 import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.context.CommandContext;
+import com.mojang.brigadier.tree.LiteralCommandNode;
 import com.scubakay.autorelog.util.Reconnect;
 import net.fabricmc.fabric.api.client.command.v2.ClientCommandManager;
 import net.fabricmc.fabric.api.client.command.v2.FabricClientCommandSource;
@@ -10,16 +11,37 @@ import net.minecraft.text.Text;
 
 public class AutoRelogCommand {
     public static void register(CommandDispatcher<FabricClientCommandSource> dispatcher, CommandRegistryAccess ignoredCommandRegistryAccess) {
-        dispatcher.register(ClientCommandManager.literal("autorelog").executes(AutoRelogCommand::run));
+        LiteralCommandNode<FabricClientCommandSource> autoRelogNode = ClientCommandManager
+                .literal("autorelog")
+                .executes(AutoRelogCommand::autorelog)
+                .build();
+
+        LiteralCommandNode<FabricClientCommandSource> cancelNode = ClientCommandManager
+                .literal("cancel")
+                .executes(AutoRelogCommand::cancel)
+                .build();
+
+        dispatcher.getRoot().addChild(autoRelogNode);
+        autoRelogNode.addChild(cancelNode);
     }
 
-    private static int run(CommandContext<FabricClientCommandSource> context) {
+    private static int autorelog(CommandContext<FabricClientCommandSource> context) {
         if (context.getSource().getClient().isInSingleplayer()) {
             context.getSource().getPlayer().sendMessage(Text.translatable("commands.autorelog_error_singleplayer"), false);
             return 1;
         }
         Reconnect.activate();
         context.getSource().getPlayer().sendMessage(Text.translatable("commands.autorelog_activated"), false);
+        return 1;
+    }
+
+    private static int cancel(CommandContext<FabricClientCommandSource> context) {
+        if (context.getSource().getClient().isInSingleplayer()) {
+            context.getSource().getPlayer().sendMessage(Text.translatable("commands.autorelog_error_singleplayer"), false);
+            return 1;
+        }
+        Reconnect.deactivate();
+        context.getSource().getPlayer().sendMessage(Text.translatable("commands.autorelog_deactivated"), false);
         return 1;
     }
 }